2016-4-19 己完成整数，浮点数，字符串，变量赋值，运算，判断，循环，输出，输入的功能

+++ json/UnUCompiler/AST/ASTIntegerNode.h
@@ -10,7 +10,7 @@
 
 namespace UnUCompiler
 {
-	class ASTIntegerNode : public ASTValueNode<int>
+	class ASTIntegerNode : public ASTValueNode<long long int>
 	{
 		friend class ASTNodeCreater;
 	public:
